Siirrä opiskelijavuositieto-raportin testit omaan tiedostoon, jätä vanhaan tiedostoon vain servletin testit
